Fix int conversions

+++ test/benchmark/benchmark_malloc_realloc.cpp
@@ -75,7 +75,7 @@ BENCHMARK_DEFINE_F(PerformanceTestFixture, benchmark_on_malloc)(
 BENCHMARK_DEFINE_F(PerformanceTestFixture, benchmark_on_realloc)(
   benchmark::State & state)
 {
-  const int malloc_size = state.range(1);
+  const int64_t malloc_size = state.range(1);
   if (malloc_size < 1) {
     state.SkipWithError("Size for allocation is too small for this test");
   }
@@ -103,7 +103,7 @@ BENCHMARK_DEFINE_F(PerformanceTestFixture, benchmark_on_realloc)(
 // with/without performance metrics
 static void malloc_args(benchmark::internal::Benchmark * b)
 {
-  for (int shift_left = 0; shift_left < 32; shift_left += 4) {
+  for (int64_t shift_left = 0; shift_left < 32; shift_left += 4) {
     b->Args({kDisablePerformanceTracking, 1 << shift_left});
     b->Args({kEnablePerformanceTracking, 1 << shift_left});
   }
@@ -117,9 +117,9 @@ BENCHMARK_REGISTER_F(PerformanceTestFixture, benchmark_on_malloc)
 // each time multiplying by 32. Each stop is tested with/without performance metrics
 static void realloc_args(benchmark::internal::Benchmark * b)
 {
-  for (int malloc_adjustment = -1; malloc_adjustment <= 1; ++malloc_adjustment) {
-    for (int realloc_shift = 0; realloc_shift < 32; realloc_shift += 8) {
-      const int malloc_shift = realloc_shift + malloc_adjustment;
+  for (int64_t malloc_adjustment = -1; malloc_adjustment <= 1; ++malloc_adjustment) {
+    for (int64_t realloc_shift = 0; realloc_shift < 32; realloc_shift += 8) {
+      const int64_t malloc_shift = realloc_shift + malloc_adjustment;
       if (malloc_shift < 0) {
         continue;
       }
